rrr: convert misc/alarm to module (rrr.alarm)
alarm.cpp is the rrr.alarm module interface unit; alarm.hpp deleted.
The Alarm class inherits FrequentJob (rrr.misc) and stores
rrr::PollThread* as a pointer-only field — full PollThread definition
not needed.

GMF forward-decl `namespace rrr { class PollThread; }` gives the field
type global-module attachment that matches reactor.h's eventual full
declaration. Without it, the pointer type would be attached to
rrr.alarm and conflict with consumers that see PollThread via
reactor.h.

Includers (2) updated to `import rrr.alarm;`:
  - src/rrr/rrr.hpp, src/rrr/misc/alock.hpp.

Measurement (rrr + rpcbench, -j32, clean):
  +stat+netinfo:  38.23 s, 175.4 MB PCM
  +alarm:         38.41 s, 175.5 MB PCM

15 modules total. 2.28× baseline.
