+++ src/rrr/misc/alock.hpp
@@ -37,9 +37,9 @@
 #include "../base/all.hpp"
 #include "../base/callback_wrapper.hpp"
 
-
-#include "alarm.hpp"
 #include "dball.hpp"
+
+import rrr.alarm;
 
 namespace rrr {
